Reorganized more things into packages, added javadoc package-infos
