+++ java/src/main/java/edu/rice/starvote/Controller.java
@@ -1,6 +1,8 @@
 package edu.rice.starvote;
 
 import edu.rice.starvote.drivers.*;
+import edu.rice.starvote.statusserver.StatusContainer;
+import edu.rice.starvote.statusserver.StatusServer;
 import edu.rice.starvote.util.GPIOListener;
 
 /**
