+++ java/src/main/java/edu/rice/starvote/ISpooler.java
@@ -10,7 +10,7 @@ public interface ISpooler {
 
     /**
      * Get the status of the paper feeder. Any status other than READY indicates that the device should not be used.
-     * A status of READY is returned only when the device is able to operate immediately.
+     * A status of READY is returned if and only if the device is able to operate immediately.
      * @return Status of paper feeder device.
      */
     DeviceStatus getStatus();
@@ -26,11 +26,11 @@ public interface ISpooler {
      *  - The device status is READY
      *
      * ##### Effects #####
-     *  - If no page is in tray, returns immediately without attempting to scan a code
-     *  - One page is accepted or ejected from the machine, if one was present
+     *  - Exactly one page is scanned. If no pages are available, do nothing and return
      *  - The ballot status is updated and sent to the status updater as appropriate
-     *  - The scanned code is sent to the validator
-     *  - The device status is set to BUSY during operation
+     *  - The scanned code is sent to `IValidator`. A valid code is accepted into the machine. An invalid code is
+     *    ejected
+     *  - The device status is always BUSY until this method returns
      *
      * ##### Postconditions #####
      *  - Device status is READY, or ERROR if error occurred
