+++ java/src/main/java/edu/rice/starvote/statusserver/StatusWebSocket.java
@@ -1,4 +1,4 @@
-package edu.rice.starvote;
+package edu.rice.starvote.statusserver;
 
 import org.eclipse.jetty.websocket.api.Session;
 import org.eclipse.jetty.websocket.api.annotations.OnWebSocketClose;
@@ -7,8 +7,6 @@ import org.eclipse.jetty.websocket.api.annotations.OnWebSocketMessage;
 import org.eclipse.jetty.websocket.api.annotations.WebSocket;
 
 import java.io.IOException;
-
-import static edu.rice.starvote.StatusServer.sessions;
 
 /**
  * WebSocket handler class for status server. Manages the global WebSocket sessions list.
@@ -20,13 +18,13 @@ public class StatusWebSocket {
 
     @OnWebSocketConnect
     public void connected(Session session) {
-        sessions.add(session);
+        StatusServer.sessions.add(session);
         System.out.println(session.getRemoteAddress().getHostString() + " connected");
     }
 
     @OnWebSocketClose
     public void closed(Session session, int statusCode, String reason) {
-        sessions.remove(session);
+        StatusServer.sessions.remove(session);
         System.out.println(session.getRemoteAddress().getHostString() + " disconnected with code " + statusCode + ": " + reason);
     }
