+++ java/src/main/java/edu/rice/starvote/statusserver/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * The status server provides the status of the ballot box (`BallotStatus`) to connected clients over HTTP and
+ * WebSockets. The client may be a simple web application that displays the ballot box status in a user-friendly
+ * manner (not included).
+ *
+ * @see edu.rice.starvote.BallotStatus
+ * @author luejerry
+ */
+package edu.rice.starvote.statusserver;
