+++ java/src/main/java/edu/rice/starvote/util/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Provides utility classes used throughout the ballot box application.
+ *
+ * @author luejerry
+ */
+package edu.rice.starvote.util;
